+++ src/extension/BuildOutputCopAnalyzer/AnalyzerAA0008.ts
@@ -1,0 +1,23 @@
+import { Analyzer } from "./Analyzer";
+import { Cops } from "./Cops";
+import { AnalyzedOutputLine } from "./Entities/AnalyzedOutputLine";
+import { AnalyzedOutputLineAA0008 } from "./Entities/AnalyzedOutputLineAA0008";
+
+export class AnalyzerAA0008 extends Analyzer {
+    analyzedLines!: AnalyzedOutputLineAA0008[];
+    
+    protected getCop(): Cops{
+        return Cops.AA0008
+    }
+    protected analyzeLineExt(analyzedOutputLine: AnalyzedOutputLine): AnalyzedOutputLineAA0008 {
+        //c:\Users\xgwsdfe\Documents\git\beat-core\app\src\codeunit\Cod5376926.FAVIncDocBarcodePDFco.al(87,25): warning AA0008: You must specify open and close parenthesis after 'AsBoolean'.
+        let regexp: RegExp = /You must specify open and close parenthesis after \'([^\']+)\'./
+        let regexpMatch: RegExpMatchArray | null = regexp.exec(analyzedOutputLine.message);
+        if (!regexpMatch)
+            throw new Error('Unexpected error.')
+        return new AnalyzedOutputLineAA0008(analyzedOutputLine, regexpMatch[1])
+    }
+    public sortDescending(): AnalyzerAA0008{
+        return super.sortDescendingImpl() as AnalyzerAA0008
+    }
+}