+++ src/extension/FixCop/IFixCop.ts
@@ -1,0 +1,4 @@
+export interface IFixCop {
+    resolve(): void
+    compilationCallback(errorLogIssues: ErrorLog.Issue[]): void
+}